refactor(platform): 收敛宿主平台条件编译至平台 crate
**🤔 背景与动机 (Why)**
- 宿主内平台差异散落多处：全屏门控、版本检测、COM 初始化、宿主侧 Windows 工具模块各自携带条件编译。
- 宿主 utils 与平台 crate 存在重复实现（环境变量展开等逐字相同），且宿主保留大量已无代码引用的 Windows 系依赖。
- 平台差异实现哲学不统一：同是全屏检测，Windows 实现在宿主内、macOS 在平台 crate。

**✨ 解决方案与影响 (What & Impact)**
- 将全屏检测、系统版本、平台就绪初始化下沉为两平台 crate 的同名函数，经平台别名层统一注入，宿主调用点不再出现条件编译。
- 删除宿主侧重复的 Windows 工具模块与无人引用的平台依赖，同步对称化平台 crate 的模块结构。
- 宿主条件编译收敛至唯一平台别名文件，后续新增平台只需在该处登记并实现同名函数。
