+++ crates/platform-windows/src/lib.rs
@@ -8,6 +8,7 @@ mod capabilities;
 mod clipboard;
 mod com_guard;
 mod focus_monitor;
+mod fullscreen;
 mod hotkey;
 mod icon;
 mod installation_monitor;
@@ -26,8 +27,9 @@ pub use app_launcher::WindowsAppLauncher;
 pub use autostart::WindowsAutoStartManager;
 pub use capabilities::windows_capabilities;
 pub use clipboard::WindowsClipboardManager;
-pub use com_guard::ComGuard;
+pub use com_guard::{init_com, ComGuard};
 pub use focus_monitor::WindowsFocusMonitor;
+pub use fullscreen::is_foreground_fullscreen;
 pub use hotkey::WindowsHotkeyManager;
 pub use icon::WindowsIconExtractor;
 pub use installation_monitor::WindowsInstallationMonitor;
@@ -39,5 +41,6 @@ pub use path_resolver::WindowsPathResolver;
 pub use resource_loader::WindowsResourceLoader;
 pub use shell::WindowsShellExecutor;
 pub use theme::{start_system_theme_monitor, WindowsThemeProvider};
+pub use utils::os_version;
 pub use window::WindowsWindowManager;
 pub use window_positioner::WindowsWindowPositioner;
