+++ src-tauri/Cargo.toml
@@ -18,39 +18,6 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 tauri-build.workspace = true
 
 [target.'cfg(target_os = "windows")'.dependencies]
-windows = { workspace = true, features = [
-    "Win32_UI_WindowsAndMessaging",
-    "Win32_Foundation",
-    "Win32_Graphics_Gdi",
-    "Management_Deployment",
-    "Win32_System_Threading",
-    "Foundation",
-    "Win32_Security",
-    "Foundation_Collections",
-    "System",
-    "Win32_System_Com",
-    "Win32_System_Com_StructuredStorage",
-    "Win32_UI_Shell",
-    "Storage_Streams",
-    "Win32_UI_Shell_PropertiesSystem",
-    "Win32_System_Variant",
-    "Win32_System_Registry",
-    "Win32_Storage_FileSystem",
-    "Win32_System_Console",
-    "Win32_System_Diagnostics",
-    "Win32_System_Diagnostics_ToolHelp",
-    "Win32_Graphics_Dwm",
-    "Win32_UI_Controls",
-    "Win32_Networking",
-    "Win32_Networking_WinInet",
-    "Win32_Globalization",
-    "Win32_System_LibraryLoader",
-    "Win32_System_Environment",
-] }
-windows-core.workspace = true
-winreg.workspace = true
-lnk.workspace = true
-uiautomation.workspace = true
 zerolaunch-platform-windows.workspace = true
 
 [target.'cfg(target_os = "macos")'.dependencies]
