+++ src-tauri/src/utils/mod.rs
@@ -3,8 +3,6 @@ pub mod font_database;
 pub mod locale;
 pub mod trace_id;
 pub mod waiting_hashmap;
-#[cfg(target_os = "windows")]
-pub mod windows;
 use chrono::{Local, NaiveDate};
 use time::OffsetDateTime;
 /// 生成当前日期的函数
